broke out util functions into separate code pieces - still want to check if theres a better import strategy
