+++ examples/20151026_SantaMonica/run_examples.sh
@@ -1,7 +1,7 @@
 # David R Thompson, Adam Erickson
 
 # Build the surface model
-python3 -c "from isofit.utils import surfmodel; surfmodel('configs/prm20151026t173213_surface_coastal.json')"
+python3 -c "from isofit.utils.surface_model import surfmodel; surfmodel('configs/prm20151026t173213_surface_coastal.json')"
 
 # Run retrievals
 isofit --level DEBUG configs/prm20151026t173213_D8W_6s.json
