+++ examples/20171108_Pasadena/run_example_modtran.sh
@@ -1,7 +1,7 @@
 # David R Thompson, Adam Erickson
 
 # Create surface model
-python3 -c "from isofit.utils.utils import surfmodel; surfmodel('configs/ang20171108t184227_surface.json')"
+python3 -c "from isofit.utils.surface_model import surfmodel; surfmodel('configs/ang20171108t184227_surface.json')"
 
 # Run experiments
 isofit --level DEBUG configs/ang20171108t173546_darklot.json
